Rename o.h to misc.h
