+++ inc/misc.h
